done second question
